refactor(core): remove the shared WorkspacePaths abstraction
Finish demoting the path model to what it is: MountFs owns the virtual
namespace, and host mapping is a real-disk detail. No caller reaches for a
"workspace path model" object anymore.

- Delete `SessionFileSystem::workspace_paths()` and the now-dead
  `ToolContext`/`SystemPromptContext::workspace_paths()` accessors. Capabilities
  use the store's `read_file`/`write_file`/`display_path`/`resolve_path`.
- Display defaults no longer go through WorkspacePaths: the trait renders the
  `/workspace` alias directly, MountFs and RealDiskFileStore override display.
- Move the host mapper (`WorkspacePaths`/`RelPath`) into RealDiskFileStore as a
  private `HostPathMap` — the one backend that maps the virtual namespace onto a
  real directory (containment, symlink rejection, worktree-root switch).
- Replace `everruns_core::workspace_paths` with a slim `session_path` module of
  host-agnostic `/workspace`-alias string helpers (`to_session_path`,
  `to_display_path`) shared by the VFS stores and the display defaults.
- Rewrite the EVE-660 conformance test to assert the same invariants through
  supported surfaces (read_file/grep/bash + on-disk reads), and update
  specs/file-store.md + specs/workspace.md.

core 2747 + integration green, runtime green, workspace builds, clippy clean.
